Fix query history and add refresh button e2e tests
- Fix query history not saving: Change Duration field from int64 to
  float64 to match the query execution response format
- Add e2e tests for schema refresh button functionality
- Change version update indicator from download icon to red dot

+++ backend/internal/storage/query_history.go
@@ -11,7 +11,7 @@ type QueryHistoryEntry struct {
 	ConnectionID   string    `json:"connectionId"`
 	ConnectionName string    `json:"connectionName"`
 	SQL            string    `json:"sql"`
-	Duration       int64     `json:"duration"`
+	Duration       float64   `json:"duration"`
 	RowCount       int       `json:"rowCount"`
 	Success        bool      `json:"success"`
 	Error          string    `json:"error,omitempty"`
